requirements.txt
BeautifulSoup==3.2.1
Django==1.6.2
MarkupSafe==0.18
Pillow==2.4.0
South==0.8.4
Unipath==1.0
amqp==1.4.3
anyjson==0.3.3
billiard==3.3.0.16
celery==3.1.9
cmsplugin-filer==0.9.5
cssmin==0.2.0
cssselect==0.9.1
distribute==0.7.3
django-allauth==0.14.1
django-appconf==0.6
django-assets==0.8
django-avatar==2.0
django-celery==3.1.9
django-celery-email==1.0.4
django-classy-tags==0.5.1
django-cms==3.0
django-compressor==1.3
-e git://github.com/comoga/django-constance@dc89ec294f67ae620ca2dc6d4019e1ecbe339917#egg=django_constance-master
django-filer==0.9.5
-e git://github.com/davedash/django-fixture-magic@5e3da5f0c587ca255ac2e508d3fb65a00f8159b0#egg=django_fixture_magic-master
django-haystack==2.1.0
django-json-field==0.5.5
django-modeltranslation==0.7.3
django-mptt==0.6.0
django-pagination==1.0.7
django-picklefield==0.3.1
django-polymorphic==0.5.4
django-postman==3.0.2
django-reversion==1.8.0
django-rosetta==0.7.4
django-sekizai==0.7
django-tinymce==1.5.2
django-widget-tweaks==1.3
djangocms-admin-style==0.2.2
djangocms-text-ckeditor==2.1.4
djangocms-column==1.3
djangocms-file==0.0.1
djangocms-flash==0.0.2
djangocms-googlemap==0.0.5
djangocms-inherit==0.0.1
djangocms-link==1.3.4
djangocms-picture==0.0.2
djangocms-style==1.3
djangocms-teaser==0.0.1
djangocms-video==0.0.1
easy-thumbnails==1.5
html5lib==0.999
jsmin==2.0.6
kombu==3.0.12
lxml==3.2.4
newrelic==2.6.0.5
oauthlib==0.6.0
polib==1.0.4
psycopg2==2.5.1
pysolr==3.1.0
python-dateutil==2.2
python-openid==2.2.5
pytz==2013.9
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
sorl-thumbnail==11.12
sqlparse==0.1.11
webassets==0.9
django-storages==1.1.8
boto==2.25.0
# comment
django-redis==3.5.2
redis==2.9.1
hiredis==0.1.4
